Exposing more Java options in flume-ng as pass throughs.
flume-ng cherrypicks arguments from its commandline to
pass through to Java.  The list includes -D* and -X*, which
are the most common, but excludes options for specifying
java agents, which are handy.  This commit adds the java
agent options to the command line.

Here's the relevant excerpt from java -help:

    -agentlib:<libname>[=<options>]
                  load native agent library <libname>, e.g. -agentlib:hprof
                  see also, -agentlib:jdwp=help and -agentlib:hprof=help
    -agentpath:<pathname>[=<options>]
                  load native agent library by full pathname
    -javaagent:<jarpath>[=<options>]
                  load Java programming language agent, see java.lang.instrument